Make a separate library for async expand methods
Pull `concurrentAsyncExpand` out from the merge extensions into a new
extension. We will add `asyncExpandSample` in a followup, so having
these in a dedicated library will make it easier for editors to find.

This could be breaking, but only for code that references the extension
name directly, which is an unlikely pattern. No usages of that pattern
exist in Dart code on Github.

Inline the implementation of `concurrentAsyncExpand` instead of
implementing a separate extension on `Stream<Stream>`
